refactor Daos in data
