+++ app/src/main/java/com/washingtondcsquad/tudee/data/localSource/daos/CategoryDao.kt
@@ -6,26 +6,23 @@ import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.Update
-import com.washingtondcsquad.tudee.data.localSource.entities.CategoryEntity
-import com.washingtondcsquad.tudee.data.utils.DataBaseConstants
+import com.washingtondcsquad.tudee.data.localSource.model.CategoryEntity
 import kotlinx.coroutines.flow.Flow
 
 @Dao
-interface DaoCategory {
-
+interface CategoryDao {
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun createCategory(category: CategoryEntity)
 
     @Delete
-    suspend fun deleteCategory(category: CategoryEntity)
+    suspend fun deleteCategory(category: CategoryEntity) //TODO delete by id
 
     @Update
     suspend fun editCategory(category: CategoryEntity)
 
-    @Query("SELECT * FROM ${DataBaseConstants.ROOM_DATABASE_NAME}")
-     fun getAllCategories(): Flow<List<CategoryEntity>>
+    @Query("SELECT * FROM ${CategoryEntity.CATEGORY_TABLE_NAME}")
+    fun getAllCategories(): Flow<List<CategoryEntity>>
 
-    @Query("SELECT * FROM ${DataBaseConstants.ROOM_DATABASE_NAME} WHERE id = :categoryId")
+    @Query("SELECT * FROM ${CategoryEntity.CATEGORY_TABLE_NAME} WHERE id = :categoryId")
     suspend fun getCategoryById(categoryId: Long): CategoryEntity
-
 }
