+++ app/src/main/java/com/washingtondcsquad/tudee/data/localSource/daos/TaskDao.kt
@@ -6,30 +6,23 @@ import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.Update
-import com.washingtondcsquad.tudee.data.localSource.entities.TaskEntity
-import com.washingtondcsquad.tudee.data.utils.DataBaseConstants
+import com.washingtondcsquad.tudee.data.localSource.model.TaskEntity
 import kotlinx.coroutines.flow.Flow
 
 @Dao
-interface DaoTask {
-
+interface TaskDao {
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun createTask(task: TaskEntity)
 
     @Delete
     suspend fun deleteTask(taskId: TaskEntity)
 
-
     @Update
     suspend fun editTask(taskId: TaskEntity)
 
-
-    @Query("SELECT * FROM ${DataBaseConstants.TASK_TABLE_NAME} WHERE id = :taskId")
+    @Query("SELECT * FROM ${TaskEntity.TASK_TABLE_NAME} WHERE id = :taskId")
     suspend fun getTaskById(taskId: Int): TaskEntity
 
-
-    @Query("SELECT * FROM ${DataBaseConstants.TASK_TABLE_NAME}")
-     fun getAllTasks(): Flow<List<TaskEntity>>
-
-
+    @Query("SELECT * FROM ${TaskEntity.TASK_TABLE_NAME}")
+    fun getAllTasks(): Flow<List<TaskEntity>>
 }
